[NUI] Add TextMapHelper class
TextUtils contains too many features and is inconsistent.
Detach the TextMapHelper functionality from the TextUtils class.
This can improves readability and reduces LOCs.
